requirements.txt
openai==1.35.0
pyyaml==6.0.1
python-dotenv==1.0.0
pydantic==2.7.0
pandas==2.2.0
openpyxl==3.1.2